// src/api/garage.d.ts
export interface components {
    schemas: {
// ...
        WorkerInfoResp: {
            /** Format: int64 */
            consecutiveErrors: number;
            /** Format: int64 */
            errors: number;
            freeform: string[];
            /** Format: int64 */
            id: number;
            lastError?: null | components["schemas"]["WorkerLastError"];
            name: string;
            /** Format: int64 */
            persistentErrors?: number | null;
            progress?: string | null;
            /** Format: int64 */
            queueLength?: number | null;
            state: components["schemas"]["WorkerStateResp"];
            /** Format: int32 */
            tranquility?: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}